Struct uic_nal::cache::Cache

pub struct Cache {
    db_conn: Connection,
}

Fields
db_conn: Connection
Implementations

impl Cache

pub fn new(nal_config: NALConfig) -> Self

Creates a new Cache instance. Try to open a database file. If file not exists - try to create it. If there already present “nal_map” table and “idx_key” in DB - does nothing. If not - creates it.

fn db_get_single_row_val(
&mut self,
key: &NodeEpKey,
column_name: &str
) -> String

Gets a single value from DB row.

pub fn node_presents(&mut self, key: &NodeEpKey) -> bool

Checks if node presents in cache.

pub fn node_get(
&mut self,
key: &NodeEpKey,
value: &mut Attributes
) -> Result<(), String>

Gets a value for the given key. If node with such key doesn’t present in DB - returns an error.

pub fn node_set(
&mut self,
key: &NodeEpKey,
attributes_new: &Attributes
) -> Result<(), String>

Updates a value for the given key. If node with such key doesn’t present in DB - creates it.

pub fn node_remove(&mut self, key: &NodeEpKey) -> Result<(), String>

Removes a node from the cache. Here is no necessity to check if node presents in cache.

pub fn get_all_nodes(&mut self) -> Result<Vec<NodeEpKey>, Error>

Returns a vector of all nodes that present in database

pub fn get_all_nodes_match_name(
&mut self,
unid: &str
) -> Result<Vec<NodeEpKey>, Error>

Returns a vector of node keys with requested UNID that presence in DB.

pub fn add_delayed_message(
&mut self,
unid: &str,
topic: &str,
payload: &str
) -> Result<(), String>

Adds a delayed message row to the nal_delayed_messages table

pub fn get_all_delayed_messages(
&mut self,
unid: &str
) -> Result<Vec<(String, String)>, Error>

Gets all delayed messages data (topic, payload) with requested UNID. Deletes all these messages trom the table.