Struct uic_nal::mqtt_handler::MqttNALHandler

pub struct MqttNALHandler<T: MqttClientTrait> {
    client: T,
    cache: Cache,
    state_matcher: TopicMatcherType,
    attributes_matcher: TopicMatcherType,
}

MqttNALHandler handles subscription messages related to Name and Location service. It subscribes to listen to the correct topics and is to be passed to the mosquitto client.

Fields
client: Tcache: Cache

Lookup cache for nodes

state_matcher: TopicMatcherType

Matches “ucl/by-unid/+/State” topic

attributes_matcher: TopicMatcherType

Matches “ucl/by-unid/+/+/#” topic

Implementations

impl<T: MqttClientTrait> MqttNALHandler<T>

pub fn new(client: T, nal_config: NALConfig) -> Result<Self, sl_status_t>

fn process_message(&mut self, msg: MosqMessage) -> Result<(), String>

fn on_state_reported(
&mut self,
topic: &str,
payload: &str
) -> Result<(), String>

Implements a "ucl/by-unid/+/State" topic handler.

Description

If payload is empty: searches for the node in cache and deletes this node and unretain it’s publications to such topics:

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Reported

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Reported

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/SupportedCommands

If payload is not empty - searches for the node in cache if found - does nothing; if not found - adds a new node to the cache and publish {"value":""} message to such topics:

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Reported

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Reported

and publish "{"value":["WriteAttributes"]}" message to ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/SupportedCommands topic

fn on_attributes_reported(
&mut self,
topic: &str,
json: JsonValue
) -> Result<(), String>

Implements a "ucl/by-unid/+/+/#" topic handler.

Description

Searches for the unid/endpoint ID “0” key in cache If node with ep_id “0” doesn’t present - this node can’t be processed

Searches for the unid/endpoint_ID key in cache If not found: create a new cache node with a given unid/endpoint_ID

If topic ends with "WriteAttributes" - parse payload If payload is Json with key either “Name” or “Location” - update node in cache If payload has another format - do nothing

If penultimate topic chunk is "LocationDescription" and topic ends with either "Desired" or "Reported" - update the desired/reported location of a given node.

For every node update operation - update the payload of retained messages for such topics:

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Reported

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Reported

The node may be “proxied” by the Protocol Controller - this means it shouldn’t update it’s “Location” attribute directly after “WriteAttributes” message received, but it has to send a message with "{"LocationDescription":"<new_location>"}" payload to ucl/by-unid/<UNID>/ep<endpoint_id>/Basic/Commands/WriteAttributes topic (that is handled by the Protocol Controller) and update the “Location” attribute value only after Protocol Controller accepts it, by publishing a message to ucl/by-unid/<UNID>/ep<endpoint_id>/Basic/Attributes/LocationDescription/<Desired/Reported>.

fn remove_nodes_via_name(&mut self, unid: &str) -> Result<(), String>

Removes all nodes with a given “unid” from cache

fn node_create(
&mut self,
key: &NodeEpKey,
attributes: &Attributes
) -> Result<(), String>

Adds a new key/value pair to the cache with a given node key/attributes values. Publishes Name and Location values to such topics (retained messages):

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Reported

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Reported Publishes the "{"value":["WriteAttributes"]}" message to topic:

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/SupportedCommands

fn node_delete(&mut self, key: &NodeEpKey) -> Result<(), String>

Deletes node with a given unid/endpoint_id key from the cache. Unretain all publications, related to this node (see unretain_all_node_messages())

fn unretain_all_node_messages(&mut self, key: &NodeEpKey) -> Result<(), String>

Unretain all publications to such topics (by publishing an empty message):

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Reported

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Reported

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/SupportedCommands

fn pub_attr_name_msgs(
&mut self,
key: &NodeEpKey,
attributes: &Attributes,
pub_des: bool,
pub_rep: bool
) -> Result<(), String>

Description

Publishes messages to such topics:

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Name/Reported

The payload is “Name” attribute value.

Arguments

pub_des - if true - publishes the “Desired” topic

pub_rep - if true - publishes the “Reported” topic

fn pub_attr_location_msgs(
&mut self,
key: &NodeEpKey,
attributes: &Attributes,
pub_des: bool,
pub_rep: bool
) -> Result<(), String>

Description

Publishes messages to such topics:

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Desired

ucl/by-unid/<UNID>/ep<endpoint_id>/NameAndLocation/Attributes/Location/Reported

The payload is “Location” attribute value.

Arguments

pub_des - if true - publishes the “Desired” topic

pub_rep - if true - publishes the “Reported” topic

fn exec_delayed_messages(&mut self, unid: &str) -> Result<(), String>

Gets all delayed messages for provided UNID and processes them - calls on_attributes_reported() function with topic and payload got from the delayed message.

fn service_stop_callback(&mut self)

Unretain all messages for all nodes.