Struct uic_nal::cache::Cache

pub struct Cache {
    db_conn: Connection,
}

Fields
db_conn: Connection
Implementations

impl Cache

pub fn new(nal_config: NALConfig) -> Self

Description

Creates a new Cache instance. Try to open a database file. If file not exists - try to create it. If tables in database not present - creates them.

Arguments

nal_config - path to database file is given from here

fn db_get_single_row_val(
&mut self,
key: &NodeEpKey,
column_name: &str
) -> String

Description

Gets a single value from a row from nal_map table.

Arguments

key - UNID and endpoint_id to search a node are given from here

column_name - the name of column in a table to get the data from

pub fn node_presents(&mut self, key: &NodeEpKey) -> bool

Description

Checks if node presents in database.

pub fn node_get(
&mut self,
key: &NodeEpKey,
value: &mut Attributes
) -> Result<(), String>

Gets a value for the given key. If node with such key doesn’t present in DB - returns an error.

pub fn node_set(
&mut self,
key: &NodeEpKey,
attributes_new: &Attributes
) -> Result<(), String>

Description

Updates attributes of node with provided key. If node with such key doesn’t present in DB - creates it.

pub fn node_remove(&mut self, key: &NodeEpKey) -> Result<(), String>

Description

Removes a node from the database. Here is no necessity to check if node presents in database.

pub fn get_all_nodes(&mut self) -> Result<Vec<NodeEpKey>, Error>

Description

Returns a vector of all nodes that present in database.

pub fn get_all_nodes_match_name(
&mut self,
unid: &str
) -> Result<Vec<NodeEpKey>, Error>

Description

Returns a vector of tuples (UNID, endpoint_id). The vector contains all nodes that present in database with provided UNID.

pub fn add_delayed_message(
&mut self,
unid: &str,
topic: &str,
payload: &str,
priority: u32
) -> Result<(), String>

Description

Adds a delayed message to nal_delayed_messages table.

pub fn get_all_delayed_messages(
&mut self,
unid: &str
) -> Result<Vec<(String, String)>, Error>

Description

Returns a vector of tuples (topic, payload). Tuples are contains delayed messages with provided UNID, sorted by priority value in ascendent order.

pub fn delete_all_delayed_messages(&mut self, unid: &str) -> Result<(), String>

Description

Deletes all delayed messages with provided UNID from delayed_messages table.

pub fn is_unid_registered(&mut self, unid: &str) -> bool

Description

Checks if a State topic has been received for specified UNID (if it was - the row with such unid value should present in unids table).

pub fn is_epidlist_received(&mut self, unid: &str) -> bool

Description

Checks if an EndpointIdList topic has been received for specified UNID (if it was - the row with such unid value should present in unids table and it’s endpoint_id_list_received value should be 1).

pub fn register_unid(&mut self, unid: &str) -> Result<(), String>

Description

Register UNID, received with a State topic. Adds a row to unids table with endpoint_id_list_received set to 0. This function should be called from the State topic handler only.

pub fn register_received_epidlist(&mut self, unid: &str) -> Result<(), String>

Description

Register receiving EndpointIdList for specified UNID. Sets endpoint_id_list_received value of row with specified UNID to 1. This function should be called from the EndpointIdList topic handler only.

pub fn remove_unid(&mut self, unid: &str) -> Result<(), String>

Description

Removes a row with specified UNID from the unids table.